rename collaborative editor context to visual editor
