+++ src/components/Collaborative/Components/Source/EditorAddSources.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useState } from "react";
 import AletheiaButton, { ButtonType } from "../../../Button";
 import { uniqueId } from "remirror";
 import SourceDialog from "../LinkToolBar/Dialog/SourceDialog";
-import { VisualEditorContext } from "../../CollaborativeEditorProvider";
+import { VisualEditorContext } from "../../VisualEditorProvider";
 import { useTranslation } from "next-i18next";
 import { PlusOutlined } from "@ant-design/icons";
 import { URL_PATTERN } from "../../hooks/useFloatingLinkState";
